Update the HSR class to adjust it to changes made to get_signal_stack()
As per df3c86d, get_signal_stack() now takes the pre-cropped PSF
template and an input argument. This change was made because cropping
the PSF is (relatively) expensive, due to the fit of the PSF center
that is required. As per the changes of this commit, the PSF template
is only cropped once at the start of the .train() method of the
HalfSiblingRegression class, and the cropped PSF template is passed to
all PixelPredictorCollection instances, which require it to construct
the forward model by calling get_signal_stack().