Refactor default experiment configuration
Move options for "psf_template" into its own section on the experiment
level of the config file, because the cropping and masking of the PSF
is now longer happening on the PixelPredictorCollection level, but is
instead done only once in the .train() method of the HSR instance.

Additionally, add "lambda_over_d" to the "dataset" section of the
configuration file, because it is essentially an instrument property.