Rename collection config to sources and refactor
After we have moved the pre-computation of the sources (based on PCA)
to the HSR class (instead of the PixelPredictorCollection) class, it
seems more intuitive to call this section of the config "sources".

Furthermore, we replace the "explained_variance_threshold" property by
a fixed number principal components to keep, as this seems to be a more
intuitive number.

The "use_forward_model" property is removed entirely. The HSR class now
implicitly determines whether or not to use forward modeling based on
the psf_template it receives (if it is None, this is interpreted as
"do not use forward modeling").