Add ExoplanetCLS model (to constrain planet coefficients to be >= 0)
This is a special case of the ConstrainedLeastSquares estimator, in the
sense that its default bounds are chosen such that the coefficient that
corresponds to the *last* feature is constrained to be non-negative.
In our HSR setting, this coefficient by convention represents the amount
of planet signal, which should not be negative, since the presence of an
exoplanet can only add more flux to a pixel, but not remove any.

(cherry picked from commit 84cbbc853a0035a7db00df6e78573ecb1b51eb86)