Change get_data_dir() to use environmental variable instead of hostname
Trying to infer the base path to the data directory based on the
hostname turned out to be problematic since the hostname of a machine
is often changed by the DHCP server of the network that said machine is
connected to. Simply setting an environmental variable HSR4HCI_DATA_DIR
and have it point to the right directory on every system this code is
supposed to run on seems like a cleaner approach.