Simplify crop_psf_template() function
Multiplying with the mask associated with a photutils.CircularAperture
object automatically crops the result to its bounding box; hence it is
not actually necessary to manually crop the PSF using Cutout2D first.

Furthermore, we can remove the cast to integer when converting the PSF
radius from units of lambda over D to pixels (since we no longer need it
to compute the crop size), which has the nice side effect that we can
scale the size of the cropped PSF template more smoothly, and not only
in integer steps.