Refactor the get_collection_residuals() method
Instead of derotating the residual frames and selecting a "cylinder" in
time from them (which is then summed up over the spatial dimensions), we
now place an aperture mask at the expected planet position onto every
residual frame, and sum up the residual flux in the aperture. Since we
no longer need to explicitly rotate any frames (which always requires
some form of interpolation), this method has the advantage that it does
not produce any Moiré artifacts in the residual stack.