Add "presubtract" option to dataset configuration
Using this keyword, we can subtract either the "mean" or the "median"
along the time axis from the stack right after we load it from the HDF
file, that is, *before* we apply any further post-processing to it.
This option may be useful, for example, to investigate if a given post-
processing method actually does more than plain median ADI.

The default value for "presubtract" is None, meaning that nothing will
be pre-subtracted from the stack after loading.