Undo 707ae2d: The -0.5 offset was correct after all (and here is why)
The offset of -0.5 is necessary, because of the following reason: The
size of the bounding box of the aperture is always odd, say 11 by 11
pixels. The center of this now would be (5.5, 5.5). However, the grid
of positions that we create by np.indices() below starts at 0, which
means the x and y positions are [0, 1, ..., 10]. The correct center
for this grid is actually (5, 5) -- hence the correction by -0.5.