Use int() instead of np.floor() to ensure result is a proper integer
np.floor() returns a numpy float, so we would need to use int() anyway
in order to make sure the result is an integer.
As a consequence, though, we can also drop the floor() call, because
int() always rounds its argument to the closest integer towards zero,
which is  identical to the behavior of floor() for positive arguments
(such as the number of apertures, which ought to be non-negative).