Do not automatically parse units when loading configuration file
Automatically parsing the units in this function leads to an
intransparent scoping behavior of the astropy.units package; it seems
safer to just to the conversion right when we need them.

Additionally, we do no longer need to carry LAMBDA_OVER_D and the
pixscale in the dataset configuration; these can now simply be taken
from the metadata attributes of the data set that we are loading.