Major refactoring of load_data() and load_default_data()
The goal of this refactoring was to change the function signatures for
loading data in a way that allows a simpler and more descriptive way of
specifying the data set in the experiment configuration.

This, however, comes at the cost of making stronger assumptions about
the location and structure of the HDF files from which we are loading
our data sets.